only work with GL textures, never pass around render functions around anymore (except when creating)
The only reason I ever started to put them everywhere a texture was needed was,
that TexBlend could need the functions to calculate the intermediate textures
needed for blending.

However, the function is not needed for this to work...

GL has several functions for manipulating textures and possibly it has even
some routines for efficiently creating a third texture by mixing two others
in a specific way, so we just give it the two texture IDs, the ratio, and have
our blend texture.

But even if not, there's still another possibility: the textures can be
recovered using glGetTexImage, blended in software, and reuploaded using
glTexImage2D. So regardless of the existance of a texture blending function
in OpenGL, I can always write my own, if necessary... therefore, no need to
carry the functions around.

Time for some code cleanup now.