TextureManager: don't break when a module refuses texture changes
Situation: the TextureManager fired off a texture_switch event to
let the rest of the program load the new texture somehow. It also
would have fired off an event that a new texture with name "Blah"
etc was loaded (only TextureManager can send it, as it's the only
one to know the names), to be displayed somehow.

Problem: Later on, Texblend might go and say, "no I won't change
the texture, because I am already doing so and it's not complete
yet."

In that case, the TextureManager would still have recorded the
texture change as successful to itself, sending off a wrong "old
texture" on the next texture change - which wouldn't be too bad
for the current code, but once Texblend is implemented, it would
blend *from* the wrong texture.

Therefore TextureManager now sends its own IDs as private fields
when it triggers a texture change, and installs a new handler that
gets run after the texture change was completed. That handler
stores the private fields into the global variables and only then
sends off the texture_changed event.

A Texblend module could just ::STOP the texture_switch handler,
then TextureManager's second handler won't get run and nothing
will be triggered or recorded.