texture switching: never return the texture_switch event, instead later fire a texture_switched event
Previous layout wouldn't have allowed texblend to properly call TextureManager's
final texture_switch handler, because if texblend invoked this event, TexBlend
would start to run just again.

So the texture_switch event is caught by whatever switching code is active
(simple or blend) and not passed on. Instead its parameters are saved, and
when the module has in any way finished changing the texture, it fires off
a texture_switched event with the exact same parameters.