TextureGenerator: properly save and restore the projection matrix stack.
During a presentation today a bug manifested whose cause was now found.
The --debug-show-planes option seemed to do nothing. It was found out that
the shader renderer of the TG resets the projection matrix but doesn't
restore it.

This bug didn't show up on my development machine because of a bug in the
radeon driver, which wrongly seems to save all matrices instead of the
current one. Therefore, the save of the texture matrix also unintentionally
saved the projection matrix, which no other GL implementation seems to do
(because the spec doesn't allow it to).