GLSL2Perl: run GLSL shaders on CPU if needed
This way, the manually provided perl versions can be removed, reducing
the need to always write two versions of a function.

The system uses some stupid regular expressions to convert well-written(tm)
GLSL (or a subset of it) to Perl, where it can be executed like the
previously provided perl subs could.