TextureGenerator: shader render: change tex memory allocation method
Obviously, and as opposed to what the specs say, it is not possible to instruct
the GL to allocate texture memory when the data pointer is 0. Also, obviously
GL_LUMINANCE for internal format doesn't work at all. So here we go, allocating
the memory manually and storing the textures internally as RGBA. Or so...