Perl2GLSL: transform perl code to GLSL, not other way round
The previously generated Perl shaders were slow like hell, which put
too many users at disadvantage. It seems to be more useful then to hand-make
the Perl shaders.

The GPU is so fast that it's unnoticeable if the shader is generated and
therefore suboptimal. But the CPU, being too slow for a task like texture
rendering anyway already, cannot be loaded with more suboptimal code.

Code is translated by parsing the Perl OP tree using the B module.

The Perl versions of the textures have been reinstated (and the last two
textures disabled, because Perl versions are missing for now).

This fixes #9 feature.