convert build system from M::B to selfwritten
because of "special needs". or so. Installing to sbin is not available
in M::B and patching it seems to be too much work or so unstable that
instead the own system will not be any less of an unclean hack than
invasively subclassing M::B. (or, in other words: "I Suck")

But at least it installs stuff properly.